pyproject.toml
[project]
name = "google_calendar_mcp"
version = "0.1.0"
description = "Google Calendar MCP stdio 서버"
authors = [
  { name = "AnByoungHyun", email = "abhyuni@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "fastmcp"
]

[project.scripts]
google_calendar_mcp = "google_calendar_mcp.server:mcp.run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
